Move `ColumnSelector` into `columns` sub-package
This creates a place for the `ColumnSchema` class to live (when it's created in an upcoming commit.)
