+++ nvtabular/__init__.py
@@ -15,7 +15,7 @@
 #
 import warnings
 
-from . import column_selector, io, workflow  # noqa
+from . import columns, io, workflow  # noqa
 from ._version import get_versions
 
 # suppress some warnings with cudf warning about column ordering with dlpack
@@ -25,7 +25,7 @@ warnings.filterwarnings("ignore", module="numba.cuda.envvars")
 
 
 WorkflowNode = workflow.WorkflowNode
-ColumnSelector = column_selector.ColumnSelector
+ColumnSelector = columns.selector.ColumnSelector
 Workflow = workflow.Workflow
 Dataset = io.dataset.Dataset
